// frontend/src/dto/TraceState.d.ts
export interface PrimitiveVal {
  readonly kind: 'PrimitiveVal'
  title: string | undefined // value displayed in a title tag (visible on hover)
  vizValue: string // value displayed as text, which might differ from original value e.g. when truncating float mantissa
  readonly primitiveValue: string
}

export interface ReferenceVal {
  readonly kind: 'ReferenceVal'
  readonly reference: number
}

export interface NullVal {
  readonly kind: 'NullVal'
}

export type Val = PrimitiveVal | ReferenceVal | NullVal

interface IVar { // interface to include these fields in all members of the union type Var
  readonly kind: 'Var' // the backend does not distinguish static vars and local vars, everything therefore has kind Var
  readonly name: string
  readonly type: string
  readonly value: Val
  changed: boolean
  heapObjectId?: number
}

export interface LocalVar extends IVar {
  class: string
  method: string
}

export interface StaticVar extends IVar {
  class: string
}

export interface LocalCondition {
  readonly id: number
  readonly expression: string
  readonly value: boolean
  readonly evaluated: boolean
}

export interface IndexExpression {
  readonly expression: string,
  readonly isVariable: boolean
}

export interface ArrayAccessTarget {
  readonly kind: 'ArrayAccessTarget'
  readonly id: number
}

export interface VariableTarget {
  readonly kind: 'VariableTarget'
  readonly name: string
}

export type AssignmentTarget = ArrayAccessTarget | VariableTarget

export interface ArrayAccess {
  readonly kind: 'ArrayAccess'
  readonly id: number
  readonly indexExpressions: IndexExpression[]
  readonly assignmentSourceVariableNames: string[]
  readonly assignmentTarget: AssignmentTarget | null,
  readonly isWrittenTo: boolean
}

export type ArrayAccessValue = {
  readonly indexValues: number[]
  readonly evaluated: boolean
  readonly arrayObjectID: number
  readonly arrayAccess: ArrayAccess
}

export interface Condition extends LocalCondition {
  readonly class: string
  readonly method: string
}

export interface HeapArrayElementVar {
  readonly kind: 'HeapArrayElementVar'
  readonly arrayId: number
  readonly type: string
  readonly value: Val
  readonly index: number
  changed: boolean
}

export type Var = LocalVar | StaticVar

interface IHeapItem {
  readonly id: number
  readonly type: string
  readonly faked: boolean
}

export interface HeapArray extends IHeapItem {
  readonly kind: 'HeapArray'
  readonly elements: HeapArrayElementVar[]
}

export interface HeapString extends IHeapItem {
  readonly kind: 'HeapString'
  string: string
  vizString: string
  readonly charArr: Var
}

export interface HeapObject extends IHeapItem {
  readonly kind: 'HeapObject'
  readonly fields: Var[]
  /** True if this object's fields were collected for specialised visualisations only and must not be shown in the Memory View. */
  readonly detailedFieldsOnly?: boolean
}

export type HeapItem = HeapArray | HeapString | HeapObject

export interface StackFrame {
  readonly kind: 'StackFrame'
  readonly line: number
  readonly class: string
  readonly method: string
  readonly signature: string
  readonly genericSignature?: string
  readonly displaySignature: string
  readonly localVariables: LocalVar[]
  readonly conditionValues: LocalCondition[]
  readonly arrayAccessValues: ArrayAccessValue[]
  readonly this?: ReferenceVal
  readonly internal: boolean
  displayText: string
}

export interface LoadedClass {
  readonly kind: 'LoadedClass'
  readonly class: string
  readonly staticFields: StaticVar[]
  readonly detailedFieldsOnly?: boolean
}
// ...
export interface ProcessedTraceState {
  readonly kind: 'ProcessedTraceState',
  readonly stateIndex: number, // the original index within the trace
  readonly localUri: string,
  readonly line: number,
  readonly heapBeforeExecution: HeapItem[],
  readonly stackBeforeExecution: StackFrame[],
  readonly loadedClassesBeforeExecution: LoadedClass[],
  readonly heapAfterExecution: HeapItem[],
  readonly stackAfterExecution: StackFrame[],
  readonly loadedClassesAfterExecution: LoadedClass[],
}